+++ tests/testdata/metadata.py
@@ -1,3 +1,5 @@
+"""Provides absolute paths to config and schema files used in tests."""
+
 from os.path import join, split
 
 path_to_testdata = split(__file__)[0]
@@ -15,3 +17,4 @@ path_to_alternative_config = join(path_to_testdata, "config/config2.yml")
 path_to_invalid_config = join(path_to_testdata, "config/config-invalid.yml")
 path_to_invalid_yml_config = join(path_to_testdata, "config/config-yml-invalid.yml")
 path_to_only_output_config = join(path_to_testdata, "config/config-only-output.yml")
+path_to_config_with_loggers = join(path_to_testdata, "config/config-with-loggers.yml")